Refactor and enhance tests for DSDBSparse and related structures
- Updated imports to reflect changes in data structure naming from DSBCOO, DSBCSR, and DSBSparse to DSDBCOO, DSDBCSR, and DSDBSparse.
- Modified test files to use the new data structure names, ensuring consistency across tests.
- Enhanced the configuration of communication settings in distributed tests to support different environments (Cupy and Numpy).
- Removed deprecated tests related to DSBSparse and adjusted the product sparsity pattern tests to focus on DSDBSparse.
- Cleaned up unnecessary comments and code in utility functions to improve readability and maintainability.
