+++ tests/datastructures/test_routines.py
@@ -6,7 +6,6 @@ from mpi4py.MPI import COMM_WORLD as global_comm
 from qttools import NDArray, sparse, xp
 from qttools.comm import comm
 from qttools.datastructures import (
-    DSBSparse,
     DSDBSparse,
     bd_matmul,
     bd_matmul_distr,
@@ -45,281 +44,121 @@ def _create_btd_coo(sizes: NDArray) -> sparse.coo_matrix:
     return sparse.coo_matrix(arr)
 
 
-def test_btd_matmul(
-    dsbsparse_type: DSBSparse,
-    block_sizes: NDArray,
-    global_stack_shape: tuple,
-):
-    """Tests the in-place addition of a DSBSparse matrix."""
-    coo = _create_btd_coo(block_sizes)
-    dsbsparse = dsbsparse_type.from_sparray(coo, block_sizes, global_stack_shape)
-    dense = dsbsparse.to_dense()
-
-    # Initalize the output matrix with the correct sparsity pattern.
-
-    out = dsbsparse_type.from_sparray(coo @ coo, block_sizes, global_stack_shape)
-
-    btd_matmul(dsbsparse, dsbsparse, out)
-
-    assert xp.allclose(dense @ dense, out.to_dense())
-
-
-def test_btd_sandwich(
-    dsbsparse_type: DSBSparse,
-    block_sizes: NDArray,
-    global_stack_shape: tuple,
-):
-    """Tests the in-place addition of a DSBSparse matrix."""
-    coo = _create_btd_coo(block_sizes)
-    dsbsparse = dsbsparse_type.from_sparray(coo, block_sizes, global_stack_shape)
-    dense = dsbsparse.to_dense()
-
-    # Initalize the output matrix with the correct sparsity pattern.
-    out = dsbsparse_type.from_sparray(coo @ coo @ coo, block_sizes, global_stack_shape)
-
-    btd_sandwich(dsbsparse, dsbsparse, out)
-
-    assert xp.allclose(dense @ dense @ dense, out.to_dense())
-
-
-def test_bd_matmul(
-    dsbsparse_type: DSBSparse,
-    block_sizes: NDArray,
-    global_stack_shape: tuple,
-):
-    """Tests the in-place addition of a DSBSparse matrix."""
-    coo = _create_btd_coo(block_sizes)
-    dsbsparse = dsbsparse_type.from_sparray(coo, block_sizes, global_stack_shape)
-    dense = dsbsparse.to_dense()
-
-    # Initalize the output matrix with the correct sparsity pattern.
-
-    out = dsbsparse_type.from_sparray(coo @ coo, block_sizes, global_stack_shape)
-
-    bd_matmul(dsbsparse, dsbsparse, out)
-
-    assert xp.allclose(dense @ dense, out.to_dense())
-
-
-def test_bd_sandwich(
-    dsbsparse_type: DSBSparse,
-    block_sizes: NDArray,
-    global_stack_shape: tuple,
-):
-    """Tests the in-place addition of a DSBSparse matrix."""
-    coo = _create_btd_coo(block_sizes)
-    dsbsparse = dsbsparse_type.from_sparray(coo, block_sizes, global_stack_shape)
-    dense = dsbsparse.to_dense()
-
-    # Initalize the output matrix with the correct sparsity pattern.
-    out = dsbsparse_type.from_sparray(coo @ coo @ coo, block_sizes, global_stack_shape)
-
-    bd_sandwich(dsbsparse, dsbsparse, out)
-
-    assert xp.allclose(dense @ dense @ dense, out.to_dense())
-
-
-def test_bd_matmul_spillover(
-    dsbsparse_type: DSBSparse,
-    block_sizes: NDArray,
-    global_stack_shape: tuple,
-):
-    """Tests the in-place addition of a DSBSparse matrix."""
-    coo = _create_btd_coo(block_sizes)
-    dsbsparse = dsbsparse_type.from_sparray(coo, block_sizes, global_stack_shape)
-    dense = dsbsparse.to_dense()
-    dense_shape = list(dense.shape)
-    NBC = 1
-    left_obc = int(sum(block_sizes[0:NBC]))
-    right_obc = int(sum(block_sizes[-NBC:]))
-    dense_shape[-2] += left_obc + right_obc
-    dense_shape[-1] += left_obc + right_obc
-
-    dense_exp = xp.zeros(tuple(dense_shape), dtype=dense.dtype)
-    dense_exp[
-        ...,
-        left_obc : left_obc + sum(block_sizes),
-        left_obc : left_obc + sum(block_sizes),
-    ] = dense
-    # simply repeat the boundaries slices
-    dense_exp[..., :left_obc, :-left_obc] = dense_exp[
-        ..., left_obc : 2 * left_obc, left_obc:
-    ]
-    dense_exp[..., :-left_obc, :left_obc] = dense_exp[
-        ..., left_obc:, left_obc : 2 * left_obc
-    ]
-    dense_exp[..., -right_obc:, right_obc:] = dense_exp[
-        ..., -2 * right_obc : -right_obc, :-right_obc
-    ]
-    dense_exp[..., right_obc:, -right_obc:] = dense_exp[
-        ..., :-right_obc, -2 * right_obc : -right_obc
-    ]
-
-    expended_product = dense_exp @ dense_exp
-    ref = expended_product[
-        ...,
-        left_obc : left_obc + sum(block_sizes),
-        left_obc : left_obc + sum(block_sizes),
-    ]
-
-    # Initalize the output matrix with the correct sparsity pattern.
-
-    out = dsbsparse_type.from_sparray(
-        sparse.coo_matrix(_get_last_2d(ref)), block_sizes, global_stack_shape
-    )
-
-    bd_matmul(dsbsparse, dsbsparse, out, spillover_correction=True)
-
-    assert xp.allclose(ref, out.to_dense())
-
-
-def test_bd_sandwich_spillover(
-    dsbsparse_type: DSBSparse,
-    block_sizes: NDArray,
-    global_stack_shape: tuple,
-):
-    """Tests the in-place addition of a DSBSparse matrix."""
-    coo = _create_btd_coo(block_sizes)
-    dsbsparse = dsbsparse_type.from_sparray(coo, block_sizes, global_stack_shape)
-    dense = dsbsparse.to_dense()
-    dense_shape = list(dense.shape)
-    NBC = 1
-    left_obc = int(sum(block_sizes[0:NBC]))
-    right_obc = int(sum(block_sizes[-NBC:]))
-    dense_shape[-2] += left_obc + right_obc
-    dense_shape[-1] += left_obc + right_obc
-    dense_exp = xp.zeros(tuple(dense_shape), dtype=dense.dtype)
-    dense_exp[
-        ...,
-        left_obc : left_obc + sum(block_sizes),
-        left_obc : left_obc + sum(block_sizes),
-    ] = dense
-    # simply repeat the boundaries slices
-    dense_exp[..., :left_obc, :-left_obc] = dense_exp[
-        ..., left_obc : 2 * left_obc, left_obc:
-    ]
-    dense_exp[..., :-left_obc, :left_obc] = dense_exp[
-        ..., left_obc:, left_obc : 2 * left_obc
-    ]
-    dense_exp[..., -right_obc:, right_obc:] = dense_exp[
-        ..., -2 * right_obc : -right_obc, :-right_obc
-    ]
-    dense_exp[..., right_obc:, -right_obc:] = dense_exp[
-        ..., :-right_obc, -2 * right_obc : -right_obc
-    ]
-
-    expended_product = dense_exp @ dense_exp @ dense_exp
-    ref = expended_product[
-        ...,
-        left_obc : left_obc + sum(block_sizes),
-        left_obc : left_obc + sum(block_sizes),
-    ]
-
-    # Initalize the output matrix with the correct sparsity pattern.
-
-    out = dsbsparse_type.from_sparray(
-        sparse.coo_matrix(_get_last_2d(ref)), block_sizes, global_stack_shape
-    )
-
-    bd_sandwich(dsbsparse, dsbsparse, out, spillover_correction=True)
-
-    assert xp.allclose(ref, out.to_dense())
-
-
-def _get_last_2d(x):
-    m, n = x.shape[-2:]
-    return x.flat[: m * n].reshape(m, n)
-
-
-@pytest.mark.mpi(min_size=3)
-class TestDistr:
-    """Tests the distributed matrix multiplication and sandwich operations."""
+class TestNotDistr:
+    """Tests the non-distributed matrix multiplication and sandwich operations."""
 
     @classmethod
     def setup_class(cls):
         """setup any state specific to the execution of the given module."""
-        _default_config = {
-            "all_to_all": "device_mpi",
-            "all_gather": "device_mpi",
-            "all_reduce": "device_mpi",
-            "bcast": "device_mpi",
-        }
+        if xp.__name__ == "cupy":
+            _default_config = {
+                "all_to_all": "host_mpi",
+                "all_gather": "host_mpi",
+                "all_reduce": "host_mpi",
+                "bcast": "host_mpi",
+            }
+        elif xp.__name__ == "numpy":
+            _default_config = {
+                "all_to_all": "device_mpi",
+                "all_gather": "device_mpi",
+                "all_reduce": "device_mpi",
+                "bcast": "device_mpi",
+            }
+
         # Configure the comm singleton.
         comm.configure(
-            block_comm_size=3,
+            block_comm_size=1,
             block_comm_config=_default_config,
             stack_comm_config=_default_config,
             override=True,
         )
 
-    def test_bd_matmul_distr(
+    def test_btd_matmul(
         self,
         dsdbsparse_type: DSDBSparse,
         block_sizes: NDArray,
         global_stack_shape: tuple,
     ):
-        """Tests the in-place addition of a DSBSparse matrix."""
+        """Tests the in-place addition of a DSDBSparse matrix."""
         coo = _create_btd_coo(block_sizes)
-        coo = global_comm.bcast(coo, root=0)
         dsdbsparse = dsdbsparse_type.from_sparray(coo, block_sizes, global_stack_shape)
         dense = dsdbsparse.to_dense()
 
         # Initalize the output matrix with the correct sparsity pattern.
 
         out = dsdbsparse_type.from_sparray(coo @ coo, block_sizes, global_stack_shape)
-        out.data[:] = 0
 
-        local_blocks, _ = get_section_sizes(len(block_sizes), comm.block.size)
-        start_block = sum(local_blocks[: comm.block.rank])
-        end_block = start_block + local_blocks[comm.block.rank]
+        btd_matmul(dsdbsparse, dsdbsparse, out)
 
-        bd_matmul_distr(
-            dsdbsparse, dsdbsparse, out, start_block=start_block, end_block=end_block
-        )
+        assert xp.allclose(dense @ dense, out.to_dense())
 
-        ref = dense @ dense
-        val = out.to_dense()
-
-        assert xp.allclose(val, ref)
-
-    def test_bd_sandwich_distr(
+    def test_btd_sandwich(
         self,
         dsdbsparse_type: DSDBSparse,
         block_sizes: NDArray,
         global_stack_shape: tuple,
     ):
-        """Tests the in-place addition of a DSBSparse matrix."""
+        """Tests the in-place addition of a DSDBSparse matrix."""
         coo = _create_btd_coo(block_sizes)
-        coo = global_comm.bcast(coo, root=0)
+        dsdbsparse = dsdbsparse_type.from_sparray(coo, block_sizes, global_stack_shape)
+        dense = dsdbsparse.to_dense()
+
+        # Initalize the output matrix with the correct sparsity pattern.
+        out = dsdbsparse_type.from_sparray(
+            coo @ coo @ coo, block_sizes, global_stack_shape
+        )
+
+        btd_sandwich(dsdbsparse, dsdbsparse, out)
+
+        assert xp.allclose(dense @ dense @ dense, out.to_dense())
+
+    def test_bd_matmul(
+        self,
+        dsdbsparse_type: DSDBSparse,
+        block_sizes: NDArray,
+        global_stack_shape: tuple,
+    ):
+        """Tests the in-place addition of a DSDBSparse matrix."""
+        coo = _create_btd_coo(block_sizes)
         dsdbsparse = dsdbsparse_type.from_sparray(coo, block_sizes, global_stack_shape)
         dense = dsdbsparse.to_dense()
 
         # Initalize the output matrix with the correct sparsity pattern.
 
-        out = dsdbsparse_type.from_sparray(
-            coo @ coo @ coo, block_sizes, global_stack_shape
-        )
-        out.data[:] = 0
+        out = dsdbsparse_type.from_sparray(coo @ coo, block_sizes, global_stack_shape)
 
-        local_blocks, _ = get_section_sizes(len(block_sizes), comm.block.size)
-        start_block = sum(local_blocks[: comm.block.rank])
-        end_block = start_block + local_blocks[comm.block.rank]
+        bd_matmul(dsdbsparse, dsdbsparse, out)
 
-        bd_sandwich_distr(
-            dsdbsparse, dsdbsparse, out, start_block=start_block, end_block=end_block
-        )
+        assert xp.allclose(dense @ dense, out.to_dense())
 
-        assert xp.allclose(dense @ dense @ dense, out.to_dense())
-
-    def test_bd_matmul_distr_spillover(
+    def test_bd_sandwich(
         self,
         dsdbsparse_type: DSDBSparse,
         block_sizes: NDArray,
         global_stack_shape: tuple,
     ):
-        """Tests the in-place addition of a DSBSparse matrix."""
+        """Tests the in-place addition of a DSDBSparse matrix."""
         coo = _create_btd_coo(block_sizes)
-        coo = global_comm.bcast(coo, root=0)
+        dsdbsparse = dsdbsparse_type.from_sparray(coo, block_sizes, global_stack_shape)
+        dense = dsdbsparse.to_dense()
+
+        # Initalize the output matrix with the correct sparsity pattern.
+        out = dsdbsparse_type.from_sparray(
+            coo @ coo @ coo, block_sizes, global_stack_shape
+        )
+
+        bd_sandwich(dsdbsparse, dsdbsparse, out)
+
+        assert xp.allclose(dense @ dense @ dense, out.to_dense())
+
+    def test_bd_matmul_spillover(
+        self,
+        dsdbsparse_type: DSDBSparse,
+        block_sizes: NDArray,
+        global_stack_shape: tuple,
+    ):
+        """Tests the in-place addition of a DSDBSparse matrix."""
+        coo = _create_btd_coo(block_sizes)
         dsdbsparse = dsdbsparse_type.from_sparray(coo, block_sizes, global_stack_shape)
         dense = dsdbsparse.to_dense()
         dense_shape = list(dense.shape)
@@ -361,32 +200,19 @@ class TestDistr:
         out = dsdbsparse_type.from_sparray(
             sparse.coo_matrix(_get_last_2d(ref)), block_sizes, global_stack_shape
         )
-        out.data[:] = 0
 
-        local_blocks, _ = get_section_sizes(len(block_sizes), comm.block.size)
-        start_block = sum(local_blocks[: comm.block.rank])
-        end_block = start_block + local_blocks[comm.block.rank]
-
-        bd_matmul_distr(
-            dsdbsparse,
-            dsdbsparse,
-            out,
-            start_block=start_block,
-            end_block=end_block,
-            spillover_correction=True,
-        )
+        bd_matmul(dsdbsparse, dsdbsparse, out, spillover_correction=True)
 
         assert xp.allclose(ref, out.to_dense())
 
-    def test_bd_sandwich_distr_spillover(
+    def test_bd_sandwich_spillover(
         self,
         dsdbsparse_type: DSDBSparse,
         block_sizes: NDArray,
         global_stack_shape: tuple,
     ):
-        """Tests the in-place addition of a DSBSparse matrix."""
+        """Tests the in-place addition of a DSDBSparse matrix."""
         coo = _create_btd_coo(block_sizes)
-        coo = global_comm.bcast(coo, root=0)
         dsdbsparse = dsdbsparse_type.from_sparray(coo, block_sizes, global_stack_shape)
         dense = dsdbsparse.to_dense()
         dense_shape = list(dense.shape)
@@ -425,6 +251,232 @@ class TestDistr:
         # Initalize the output matrix with the correct sparsity pattern.
 
         out = dsdbsparse_type.from_sparray(
+            sparse.coo_matrix(_get_last_2d(ref)), block_sizes, global_stack_shape
+        )
+
+        bd_sandwich(dsdbsparse, dsdbsparse, out, spillover_correction=True)
+
+        assert xp.allclose(ref, out.to_dense())
+
+
+def _get_last_2d(x):
+    m, n = x.shape[-2:]
+    return x.flat[: m * n].reshape(m, n)
+
+
+@pytest.mark.mpi(min_size=3)
+class TestDistr:
+    """Tests the distributed matrix multiplication and sandwich operations."""
+
+    @classmethod
+    def setup_class(cls):
+        """setup any state specific to the execution of the given module."""
+        if xp.__name__ == "cupy":
+            _default_config = {
+                "all_to_all": "host_mpi",
+                "all_gather": "host_mpi",
+                "all_reduce": "host_mpi",
+                "bcast": "host_mpi",
+            }
+        elif xp.__name__ == "numpy":
+            _default_config = {
+                "all_to_all": "device_mpi",
+                "all_gather": "device_mpi",
+                "all_reduce": "device_mpi",
+                "bcast": "device_mpi",
+            }
+
+        # Configure the comm singleton.
+        comm.configure(
+            block_comm_size=3,
+            block_comm_config=_default_config,
+            stack_comm_config=_default_config,
+            override=True,
+        )
+
+    def test_bd_matmul_distr(
+        self,
+        dsdbsparse_type_dist: DSDBSparse,
+        block_sizes: NDArray,
+        global_stack_shape: tuple,
+    ):
+        """Tests the in-place addition of a DSDBSparse matrix."""
+        coo = _create_btd_coo(block_sizes)
+        coo = global_comm.bcast(coo, root=0)
+        dsdbsparse = dsdbsparse_type_dist.from_sparray(
+            coo, block_sizes, global_stack_shape
+        )
+        dense = dsdbsparse.to_dense()
+
+        # Initalize the output matrix with the correct sparsity pattern.
+
+        out = dsdbsparse_type_dist.from_sparray(
+            coo @ coo, block_sizes, global_stack_shape
+        )
+        out.data[:] = 0
+
+        local_blocks, _ = get_section_sizes(len(block_sizes), comm.block.size)
+        start_block = sum(local_blocks[: comm.block.rank])
+        end_block = start_block + local_blocks[comm.block.rank]
+
+        bd_matmul_distr(
+            dsdbsparse, dsdbsparse, out, start_block=start_block, end_block=end_block
+        )
+
+        ref = dense @ dense
+        val = out.to_dense()
+
+        assert xp.allclose(val, ref)
+
+    def test_bd_sandwich_distr(
+        self,
+        dsdbsparse_type_dist: DSDBSparse,
+        block_sizes: NDArray,
+        global_stack_shape: tuple,
+    ):
+        """Tests the in-place addition of a DSDBSparse matrix."""
+        coo = _create_btd_coo(block_sizes)
+        coo = global_comm.bcast(coo, root=0)
+        dsdbsparse = dsdbsparse_type_dist.from_sparray(
+            coo, block_sizes, global_stack_shape
+        )
+        dense = dsdbsparse.to_dense()
+
+        # Initalize the output matrix with the correct sparsity pattern.
+
+        out = dsdbsparse_type_dist.from_sparray(
+            coo @ coo @ coo, block_sizes, global_stack_shape
+        )
+        out.data[:] = 0
+
+        local_blocks, _ = get_section_sizes(len(block_sizes), comm.block.size)
+        start_block = sum(local_blocks[: comm.block.rank])
+        end_block = start_block + local_blocks[comm.block.rank]
+
+        bd_sandwich_distr(
+            dsdbsparse, dsdbsparse, out, start_block=start_block, end_block=end_block
+        )
+
+        assert xp.allclose(dense @ dense @ dense, out.to_dense())
+
+    def test_bd_matmul_distr_spillover(
+        self,
+        dsdbsparse_type_dist: DSDBSparse,
+        block_sizes: NDArray,
+        global_stack_shape: tuple,
+    ):
+        """Tests the in-place addition of a DSDBSparse matrix."""
+        coo = _create_btd_coo(block_sizes)
+        coo = global_comm.bcast(coo, root=0)
+        dsdbsparse = dsdbsparse_type_dist.from_sparray(
+            coo, block_sizes, global_stack_shape
+        )
+        dense = dsdbsparse.to_dense()
+        dense_shape = list(dense.shape)
+        NBC = 1
+        left_obc = int(sum(block_sizes[0:NBC]))
+        right_obc = int(sum(block_sizes[-NBC:]))
+        dense_shape[-2] += left_obc + right_obc
+        dense_shape[-1] += left_obc + right_obc
+
+        dense_exp = xp.zeros(tuple(dense_shape), dtype=dense.dtype)
+        dense_exp[
+            ...,
+            left_obc : left_obc + sum(block_sizes),
+            left_obc : left_obc + sum(block_sizes),
+        ] = dense
+        # simply repeat the boundaries slices
+        dense_exp[..., :left_obc, :-left_obc] = dense_exp[
+            ..., left_obc : 2 * left_obc, left_obc:
+        ]
+        dense_exp[..., :-left_obc, :left_obc] = dense_exp[
+            ..., left_obc:, left_obc : 2 * left_obc
+        ]
+        dense_exp[..., -right_obc:, right_obc:] = dense_exp[
+            ..., -2 * right_obc : -right_obc, :-right_obc
+        ]
+        dense_exp[..., right_obc:, -right_obc:] = dense_exp[
+            ..., :-right_obc, -2 * right_obc : -right_obc
+        ]
+
+        expended_product = dense_exp @ dense_exp
+        ref = expended_product[
+            ...,
+            left_obc : left_obc + sum(block_sizes),
+            left_obc : left_obc + sum(block_sizes),
+        ]
+
+        # Initalize the output matrix with the correct sparsity pattern.
+
+        out = dsdbsparse_type_dist.from_sparray(
+            sparse.coo_matrix(_get_last_2d(ref)), block_sizes, global_stack_shape
+        )
+        out.data[:] = 0
+
+        local_blocks, _ = get_section_sizes(len(block_sizes), comm.block.size)
+        start_block = sum(local_blocks[: comm.block.rank])
+        end_block = start_block + local_blocks[comm.block.rank]
+
+        bd_matmul_distr(
+            dsdbsparse,
+            dsdbsparse,
+            out,
+            start_block=start_block,
+            end_block=end_block,
+            spillover_correction=True,
+        )
+
+        assert xp.allclose(ref, out.to_dense())
+
+    def test_bd_sandwich_distr_spillover(
+        self,
+        dsdbsparse_type_dist: DSDBSparse,
+        block_sizes: NDArray,
+        global_stack_shape: tuple,
+    ):
+        """Tests the in-place addition of a DSDBSparse matrix."""
+        coo = _create_btd_coo(block_sizes)
+        coo = global_comm.bcast(coo, root=0)
+        dsdbsparse = dsdbsparse_type_dist.from_sparray(
+            coo, block_sizes, global_stack_shape
+        )
+        dense = dsdbsparse.to_dense()
+        dense_shape = list(dense.shape)
+        NBC = 1
+        left_obc = int(sum(block_sizes[0:NBC]))
+        right_obc = int(sum(block_sizes[-NBC:]))
+        dense_shape[-2] += left_obc + right_obc
+        dense_shape[-1] += left_obc + right_obc
+        dense_exp = xp.zeros(tuple(dense_shape), dtype=dense.dtype)
+        dense_exp[
+            ...,
+            left_obc : left_obc + sum(block_sizes),
+            left_obc : left_obc + sum(block_sizes),
+        ] = dense
+        # simply repeat the boundaries slices
+        dense_exp[..., :left_obc, :-left_obc] = dense_exp[
+            ..., left_obc : 2 * left_obc, left_obc:
+        ]
+        dense_exp[..., :-left_obc, :left_obc] = dense_exp[
+            ..., left_obc:, left_obc : 2 * left_obc
+        ]
+        dense_exp[..., -right_obc:, right_obc:] = dense_exp[
+            ..., -2 * right_obc : -right_obc, :-right_obc
+        ]
+        dense_exp[..., right_obc:, -right_obc:] = dense_exp[
+            ..., :-right_obc, -2 * right_obc : -right_obc
+        ]
+
+        expended_product = dense_exp @ dense_exp @ dense_exp
+        ref = expended_product[
+            ...,
+            left_obc : left_obc + sum(block_sizes),
+            left_obc : left_obc + sum(block_sizes),
+        ]
+
+        # Initalize the output matrix with the correct sparsity pattern.
+
+        out = dsdbsparse_type_dist.from_sparray(
             sparse.coo_matrix(_get_last_2d(ref)), block_sizes, global_stack_shape
         )
         out.data[:] = 0
