+++ tests/greens_function_solver/test_selected_inv.py
@@ -1,14 +1,14 @@
 # Copyright (c) 2024 ETH Zurich and the authors of the qttools package.
 
 from qttools import NDArray, sparse, xp
-from qttools.datastructures import DSBSparse
+from qttools.datastructures import DSDBSparse
 from qttools.greens_function_solver import GFSolver
 
 
 def test_selected_inv(
     bt_dense: NDArray,
     gfsolver_type: GFSolver,
-    dsbsparse_type: DSBSparse,
+    dsdbsparse_type: DSDBSparse,
     out: bool,
     max_batch_size: int,
     block_sizes: NDArray,
@@ -21,15 +21,15 @@ def test_selected_inv(
     coo = sparse.coo_matrix(bt_dense)
 
     block_sizes = block_sizes
-    dsbsparse = dsbsparse_type.from_sparray(coo, block_sizes, global_stack_shape)
+    dsdbsparse = dsdbsparse_type.from_sparray(coo, block_sizes, global_stack_shape)
 
     solver = gfsolver_type(max_batch_size=max_batch_size)
 
     if out:
-        gf_inv = dsbsparse_type.zeros_like(dsbsparse)
-        solver.selected_inv(dsbsparse, out=gf_inv)
+        gf_inv = dsdbsparse_type.zeros_like(dsdbsparse)
+        solver.selected_inv(dsdbsparse, out=gf_inv)
     else:
-        gf_inv = solver.selected_inv(dsbsparse)
+        gf_inv = solver.selected_inv(dsdbsparse)
 
     bt_mask_broadcasted = xp.broadcast_to(
         bt_mask, (*global_stack_shape, *bt_mask.shape)
