+++ tests/obc/test_spectral.py
@@ -3,7 +3,7 @@
 import pytest
 
 from qttools import NDArray, xp
-from qttools.datastructures.dsbsparse import _block_view
+from qttools.datastructures.dsdbsparse import _block_view
 from qttools.nevp import NEVP
 from qttools.obc import OBCMemoizer, Spectral
